Fix & Add: Booking Login, Add type, Refactoring Login code
