+++ src/app/booking/Components/Info/BookPaymentMethod.tsx
@@ -15,13 +15,11 @@ import { ReservationProgress } from "@shared/components/ReservationProgress/Rese
 import { Reservation } from "@/@type/Type";
 
 export const BookPaymentMethod = ({
-  paySelect,
   // onPaySelectHandle,
   checkState,
   checkHandled,
   reservation,
 }: {
-  paySelect: string;
   // onPaySelectHandle: ChangeEvent<{ value: unknown }>;
   checkState: boolean;
   checkHandled: (e: ChangeEvent<HTMLInputElement>) => void;
@@ -31,22 +29,20 @@ export const BookPaymentMethod = ({
     <>
       <SecondHead className="mt-4">결제 수단</SecondHead>
       {/* 카드 번호, 유효기간, vs */}
-      <Horizon />
       <div className="mt-2 ml-4">
-        <NormalText>예약 확정 전에는 요금이 청구되지 않습니다.</NormalText>
+        <NormalText>
+          안전 계좌에 입금 후 예약 확정이 진행됩니다. 또한 사기 방지를 위해 입주
+          이후 호스트에게 송급됩니다.
+        </NormalText>
         <Box className="mt-4 mx-9" justifyContent="center" alignItems="center">
-          {/* <BookPaymentForm
-            paySelect={paySelect}
-            onPaySelectHandle={onPaySelectHandle}
-          /> */}
           <ReservationProgress reservation={reservation} hostPosition={false} />
-          {/* {paySelect === "account" ? <AccountPayment /> : <PaymentForm />} */}
           <NormalText className="mt-2">
             <Checkbox checked={checkState} onChange={checkHandled} />
             체크박스를 클릭하시면 이체를 완료하셨음을 동의하는 것입니다.
           </NormalText>
         </Box>
       </div>
+      <Horizon />
     </>
   );
 };
