+++ src/app/booking/page.tsx
@@ -11,10 +11,16 @@ import { ChangeEvent, useState } from "react";
 import { FetchReservationPost } from "@shared/components/FetchList/FetchList";
 import { Reservation } from "@/@type/Type";
 import { bookingPopUpStore } from "@store/BookingPopUpStore";
+import {
+  DisableButton,
+  NormalButton,
+  Wrapper,
+} from "@shared/styles/Public.styles";
+import { useUserInfoStore } from "@store/UserInfoStore";
+import { LoginContent } from "@shared/components/loginComponents/LoginContent";
 
 export default function Booking(userId: string) {
   // useTitle('예약하기 | ItHome');
-
   const { reservation, postKey } = bookingPopUpStore((state) => ({
     reservation: state.reservation,
     postKey: state.postKey,
@@ -38,40 +44,82 @@ export default function Booking(userId: string) {
   const totalRefundDate = CalulateDate(startDay, -7);
   const partRefundDate = CalulateDate(startDay, -3);
 
-  const [paySelect, setPaySelect] = useState("account");
-  // const onPaySelectHandle = (e: ChangeEvent<{ value: unknown }>) => {
-  //   setPaySelect(e.target.value as string);
-  // };
   const [checkState, setCheckState] = useState(false);
 
   const checkHandled = (e: ChangeEvent<HTMLInputElement>) => {
     setCheckState(!checkState);
   };
+  const { userInfo } = useUserInfoStore(); // 로그인 테스트 (true: 로그인, false: 로그아웃)
 
   const totalDay = getDateDiff(startDay, endDay);
   const monthPay = reservation.pay * 28;
   return (
-    <div className="ml-4 w-4/5 items-center">
-      <BookPriceAndDate
-        startDay={startDay}
-        endDay={endDay}
-        totalDay={totalDay}
-        totalPay={totalPay}
-        monthPay={monthPay}
-      />
-      <BookPaymentMethod
-        paySelect={paySelect}
-        reservation={reservation}
-        // onPaySelectHandle={onPaySelectHandle}
-        checkState={checkState}
-        checkHandled={checkHandled}
-      />
-      <BookRefundRule
-        totalRefundDate={totalRefundDate}
-        partRefundDate={partRefundDate}
-        checkState={checkState}
-        handlePostReservation={handlePostReservation}
-      />
-    </div>
+    <Wrapper>
+      <div
+        style={{
+          margin: "0px 80px 0px 80px",
+        }}
+        className="items-center"
+      >
+        <div
+          style={{
+            margin: "20px 0px 20px 0px",
+          }}
+        >
+          <BookPriceAndDate
+            startDay={startDay}
+            endDay={endDay}
+            totalDay={totalDay}
+            totalPay={totalPay}
+            monthPay={monthPay}
+          />
+        </div>
+        <div
+          style={{
+            margin: "20px 0px 20px 0px",
+          }}
+        >
+          <BookPaymentMethod
+            reservation={reservation}
+            // onPaySelectHandle={onPaySelectHandle}
+            checkState={checkState}
+            checkHandled={checkHandled}
+          />
+        </div>
+
+        <div
+          style={{
+            margin: "20px 0px 20px 0px",
+          }}
+        >
+          <BookRefundRule
+            totalRefundDate={totalRefundDate}
+            partRefundDate={partRefundDate}
+          />
+        </div>
+
+        <div
+          style={{
+            margin: "20px 0px 20px 0px",
+          }}
+        >
+          {userInfo.user_id ? (
+            <>
+              {checkState ? (
+                <NormalButton onClick={handlePostReservation}>
+                  예약 요청하기
+                </NormalButton>
+              ) : (
+                <DisableButton disabled>예약 요청하기</DisableButton>
+              )}
+            </>
+          ) : (
+            <>
+              <LoginContent setPopUpState={undefined} />
+            </>
+          )}
+        </div>
+      </div>
+    </Wrapper>
   );
 }
